huge change from plot, remove "ngoai_ngu" comb and add plot to get 6 plots
