+++ plot/freqSub.py
@@ -1,90 +1,68 @@
 import pandas as pd
-import numpy as np
+import seaborn as sns
 import matplotlib.pyplot as plt
 
-# Đọc file CSV
 data2023 = pd.read_csv("diem2023.csv")
 data2024 = pd.read_csv("diem2024.csv")
 
 
-def plot_percentage_histogram(data1, data2, subject, year1, year2):
+def freqSub(data1, data2, year1, year2):
     """
-    Vẽ biểu đồ histogram để so sánh phân bố điểm của một môn học giữa hai năm,
-    đồng thời hiển thị tỷ lệ phần trăm của mỗi khoảng điểm và các percentiles.
+    Vẽ biểu đồ histogram so sánh phân bố điểm của các môn học giữa hai năm.
 
     Args:
         data1 (pd.DataFrame): Dữ liệu điểm thi của năm thứ nhất, với mỗi cột là một môn học và các hàng là điểm của từng học sinh.
         data2 (pd.DataFrame): Dữ liệu điểm thi của năm thứ hai, với mỗi cột là một môn học và các hàng là điểm của từng học sinh.
-        subject (str): Tên môn học cần vẽ biểu đồ (ví dụ: 'toan').
         year1 (int): Năm học của dữ liệu thứ nhất (ví dụ: 2023).
         year2 (int): Năm học của dữ liệu thứ hai (ví dụ: 2024).
 
     Returns:
-        None: Hiển thị trực tiếp biểu đồ histogram so sánh phân bố điểm của một môn học giữa hai năm.
+        None: Hiển thị trực tiếp biểu đồ histogram so sánh phân bố điểm của các môn học giữa hai năm.
 
     Raises:
-        ValueError: Nếu môn học không có trong dữ liệu của một trong hai năm.
+        None
     """
-    # Kiểm tra xem môn học có trong dữ liệu không
-    if subject not in data1.columns or subject not in data2.columns:
-        print(f"Môn học {subject} không có trong dữ liệu.")
-        return
+    # Liệt kê các môn học và tên tiếng Việt tương ứng
+    subjects = {
+        "toan": "Toán",
+        "ngu_van": "Ngữ văn",
+        "ngoai_ngu": "Ngoại ngữ",
+        "vat_li": "Vật lý",
+        "hoa_hoc": "Hóa học",
+        "sinh_hoc": "Sinh học",
+        "lich_su": "Lịch sử",
+        "dia_li": "Địa lý",
+        "gdcd": "GDCD",
+    }
 
-    # Tạo biểu đồ
-    fig, axs = plt.subplots(1, 2, figsize=(18, 6), sharey=True)
+    # Tạo DataFrame mới chứa điểm và năm
+    data1["Year"] = year1
+    data2["Year"] = year2
+    combined_data = pd.concat([data1, data2], ignore_index=True)
 
-    # Vẽ histogram và tính tỷ lệ phần trăm cho năm 2023
-    counts1, bins1, patches1 = axs[0].hist(
-        data1[subject].dropna(),
-        bins=10,
-        edgecolor="black",
-        color="skyblue",
-        density=True,
-    )
-    counts1 = counts1 * (bins1[1] - bins1[0]) * 100  # Chuyển đổi tỷ lệ phần trăm
-    for count, patch in zip(counts1, patches1):
-        height = patch.get_height()
-        axs[0].text(
-            patch.get_x() + patch.get_width() / 2.0,
-            height,
-            f"{count:.1f}%",
-            ha="center",
-            va="bottom",
+    # Vẽ biểu đồ histogram cho từng môn
+    fig, axes = plt.subplots(3, 3, figsize=(18, 12))
+    fig.suptitle("Phân Bố Điểm Các Môn THPTQG 2 năm 2023, 2024 của TPHCM", fontsize=16, y=1)
+
+    for ax, (subject, vietnamese_name) in zip(axes.flatten(), subjects.items()):
+        sns.histplot(
+            data=combined_data,
+            x=subject,
+            hue="Year",
+            bins=20,
+            kde=False,
+            ax=ax,
+            palette="pastel",
+            alpha=0.6,
         )
-    axs[0].set_title(f"Năm {year1}")
-    axs[0].set_xlabel("Điểm")
-    axs[0].set_ylabel("Tần suất (%)")
+        ax.set_title(vietnamese_name)
+        ax.set_xlabel("Điểm")
+        ax.set_ylabel("Tần suất")
+        ax.grid(True)
 
-    # Vẽ histogram và tính tỷ lệ phần trăm cho năm 2024
-    counts2, bins2, patches2 = axs[1].hist(
-        data2[subject].dropna(),
-        bins=10,
-        edgecolor="black",
-        color="lightcoral",
-        density=True,
-    )
-    counts2 = counts2 * (bins2[1] - bins2[0]) * 100  # Chuyển đổi tỷ lệ phần trăm
-    for count, patch in zip(counts2, patches2):
-        height = patch.get_height()
-        axs[1].text(
-            patch.get_x() + patch.get_width() / 2.0,
-            height,
-            f"{count:.1f}%",
-            ha="center",
-            va="bottom",
-        )
-    axs[1].set_title(f"Năm {year2}")
-    axs[1].set_xlabel("Điểm")
-
-    # Thêm tiêu đề tổng thể cho hình
-    plt.suptitle(
-        f"So Sánh Phân Bố Điểm Môn {subject.capitalize()} Giữa Các Năm",
-        fontsize=16,
-        y=1,
-    )
     plt.tight_layout(pad=3.0)
     plt.show()
 
 
-# Gọi hàm để vẽ biểu đồ Percentage Histogram cho một môn, ví dụ: Toán
-plot_percentage_histogram(data2023, data2024, "toan", 2023, 2024)
+# Gọi hàm với dữ liệu của năm 2023 và 2024
+freqSub(data2023, data2024, 2023, 2024)
